requirements.txt
pandas
#tensorflow
facenet-pytorch
scikit-learn
retinaface-pytorch